[ACA-4728] further reduce code duplication, remove/replace faulty unit tests
